Shared/requirements.txt
# Core DeepAgents Framework
deepagents>=0.1.0

# LangChain and LangGraph
langgraph>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5

# Search API (for Deep Research agent)
tavily-python>=0.3.0

# Model Context Protocol
langchain-mcp-adapters>=0.1.0

# Utilities
python-dotenv>=1.0.0

# Note: open_deep_research files are copied into src/open_deep_research/
# We removed langchain-anthropic since we're only using OpenAI GPT models
